backend/API/requirement.txt
flask
google-generativeai
openai
